changed everything to double

+++ ML-Models/k_means_clustering.cpp
@@ -10,22 +10,22 @@
 class KMeans {
 public:
     int k;
-    std::map<int, std::vector<std::vector<float>>> cluster_assignments;
-    std::map<int, std::vector<float>> cluster_coordinates;
+    std::map<int, std::vector<std::vector<double>>> cluster_assignments;
+    std::map<int, std::vector<double>> cluster_coordinates;
 
     KMeans(int k) {
         this->k = k;
     }
 
-    std::map<int, std::vector<std::vector<float>>> fit(std::vector<std::vector<float>> data_points_vector) {
+    std::map<int, std::vector<std::vector<double>>> fit(std::vector<std::vector<double>> data_points_vector) {
         std::random_device rd{};
         std::mt19937 gen{rd()};
         std::normal_distribution<double> distribution{0.0, 1.0};
 
         int num_features = data_points_vector[0].size();
-        std::vector<float> feature_averages(num_features);
+        std::vector<double> feature_averages(num_features);
 
-        std::vector<std::vector<float>> data_points_vector_T = takeTranspose(data_points_vector);
+        std::vector<std::vector<double>> data_points_vector_T = takeTranspose(data_points_vector);
 
         for (int i = 0; i < num_features; i++) {
             int sum = std::accumulate(data_points_vector_T[i].begin(), data_points_vector_T[i].end(), 0);
@@ -33,7 +33,7 @@ public:
         }
 
         for (int i = 0; i < k; i++) {
-            std::vector<float> cluster_coords(num_features);
+            std::vector<double> cluster_coords(num_features);
             cluster_coordinates.insert({i, cluster_coords});
 
             for (int j = 0; j < cluster_coords.size(); j++) {
@@ -43,14 +43,14 @@ public:
 
         // Assign each data point to the closest cluster
         for (int i = 0; i < data_points_vector.size(); i++) {
-            std::vector<float> data_coord = data_points_vector[i];
+            std::vector<double> data_coord = data_points_vector[i];
 
             int closest_cluster = 0;
-            float dist_closest_cluster = std::numeric_limits<int>::max();
+            double dist_closest_cluster = std::numeric_limits<int>::max();
 
             for (int j = 0; j < k; j++) {
-                std::vector<float> cluster_coord = cluster_coordinates[j];
-                float dist = calculateNorm(subtractVectors(cluster_coord, data_coord));
+                std::vector<double> cluster_coord = cluster_coordinates[j];
+                double dist = calculateNorm(subtractVectors(cluster_coord, data_coord));
 
                 if (dist < dist_closest_cluster) {
                     closest_cluster = j;
@@ -64,15 +64,15 @@ public:
         // Iteratively refine the cluster centroids
         bool optimized = false;
         while (!optimized) {
-            float cluster_coord_cum_delta = 0;
+            double cluster_coord_cum_delta = 0;
             // Calculate new cluster centroids
             for (int i = 0; i < k; i++) {
                 if (cluster_assignments[i].empty()) {
                 continue; 
                 }
-                std::vector<float> old_cluster_coords = cluster_coordinates[i];
+                std::vector<double> old_cluster_coords = cluster_coordinates[i];
                 for (int j = 0; j < num_features; j++) {
-                    std::vector<float> feature_vals = takeTranspose(cluster_assignments[i])[j];
+                    std::vector<double> feature_vals = takeTranspose(cluster_assignments[i])[j];
                     int sum = std::accumulate(
                         feature_vals.begin(),
                         feature_vals.end(),
@@ -80,22 +80,22 @@ public:
                     );
                     cluster_coordinates[i][j] = sum / cluster_assignments[i].size();
                 }
-                std::vector<float> new_cluster_coords = cluster_coordinates[i];
+                std::vector<double> new_cluster_coords = cluster_coordinates[i];
                 cluster_coord_cum_delta += calculateNorm(subtractVectors(old_cluster_coords, new_cluster_coords));
             }
-            float cluster_coord_avg_delta = cluster_coord_cum_delta / k;
+            double cluster_coord_avg_delta = cluster_coord_cum_delta / k;
 
             // Reassign points to the closest cluster
             cluster_assignments.clear();
             for (int i = 0; i < data_points_vector.size(); i++) {
-                std::vector<float> data_coord = data_points_vector[i];
+                std::vector<double> data_coord = data_points_vector[i];
 
                 int closest_cluster = 0;
-                float dist_closest_cluster = std::numeric_limits<int>::max();
+                double dist_closest_cluster = std::numeric_limits<int>::max();
 
                 for (int j = 0; j < k; j++) {
-                    std::vector<float> cluster_coord = cluster_coordinates[j];
-                    float dist = calculateNorm(subtractVectors(cluster_coord, data_coord));
+                    std::vector<double> cluster_coord = cluster_coordinates[j];
+                    double dist = calculateNorm(subtractVectors(cluster_coord, data_coord));
 
                     if (dist < dist_closest_cluster) {
                         closest_cluster = j;
