+++ libs/Data-Processing/ReadCSV.h
@@ -5,13 +5,13 @@
 #include <fstream>
 #include <sstream>
 
-std::vector<std::vector<float>> parseCSV(std::string file_name);
+std::vector<std::vector<double>> parseCSV(std::string file_name);
 std::vector<std::string> getCSVHeaders(std::string file_name);
 std::vector<std::string> getCSVIndices(std::string file_name, int index_col);
 void toCSV(const std::string &file_name,
-           const std::vector<std::vector<float>> &data,
-           const std::vector<float> &labels,
-           const std::vector<float> &predictions);
+           const std::vector<std::vector<double>> &data,
+           const std::vector<double> &labels,
+           const std::vector<double> &predictions);
 
 
 #endif
