+++ libs/Math/GenFunctions.cpp
@@ -4,19 +4,19 @@
 #include <iostream>
 #include "Output.h"
 
-float calculateMSE(std::vector<float> predictions, std::vector<float> labels) {
+double calculateMSE(std::vector<double> predictions, std::vector<double> labels) {
     int num_elements = predictions.size();
 
-    std::vector<float> resultant = subtractVectors(labels, predictions);
+    std::vector<double> resultant = subtractVectors(labels, predictions);
     return innerProduct(resultant, resultant) / num_elements;
 }
 
 
-float calculateLogLoss(std::vector<float> predictions, std::vector<float> labels) {
-    float cumSum = 0;
-    const float epsilon = 1e-10;
+double calculateLogLoss(std::vector<double> predictions, std::vector<double> labels) {
+    double cumSum = 0;
+    const double epsilon = 1e-10;
     for (int i = 0; i < predictions.size(); i++) {
-        float clipped_prediction = std::max(epsilon, std::min(1-epsilon, predictions[i]));
+        double clipped_prediction = std::max(epsilon, std::min(1-epsilon, predictions[i]));
         cumSum += (labels[i] * std::log(clipped_prediction)) + ((1-labels[i]) * std::log((1-clipped_prediction)));
     }
     return (-1 * (cumSum / predictions.size()));
@@ -24,8 +24,8 @@ float calculateLogLoss(std::vector<float> predictions, std::vector<float> labels
 
 
 
-std::vector<float> thresholdFunction(std::vector<float> softPredictions, float threshhold) {
-    std::vector<float> hardPredictions(softPredictions.size());
+std::vector<double> thresholdFunction(std::vector<double> softPredictions, double threshhold) {
+    std::vector<double> hardPredictions(softPredictions.size());
 
     for (int i = 0; i < softPredictions.size(); i++) {
         if (softPredictions[i] >= threshhold) {
@@ -38,46 +38,46 @@ std::vector<float> thresholdFunction(std::vector<float> softPredictions, float t
     return hardPredictions;
 }
 
-float sigmoid(float value) {
+double sigmoid(double value) {
     if (value >= 0) {
         return 1.0 / (1.0 + std::exp(-value));
     } else {
-        float exp_val = std::exp(value);
+        double exp_val = std::exp(value);
         return exp_val / (1.0 + exp_val);
     }
 }
 
-float calculateMean(std::vector<float> v1) {
+double calculateMean(std::vector<double> v1) {
     return (accumulateVector(v1) / v1.size());
 }
 
-float calculateSTD(std::vector<float> v1) {
+double calculateSTD(std::vector<double> v1) {
     int num_elements = v1.size();
-    float mean = calculateMean(v1);
+    double mean = calculateMean(v1);
 
-    std::vector<float> mean_vector = createVector(mean, num_elements);
+    std::vector<double> mean_vector = createVector(mean, num_elements);
 
-    std::vector<float> normalized = subtractVectors(v1, mean_vector);
-    float IP = innerProduct(normalized, normalized);
+    std::vector<double> normalized = subtractVectors(v1, mean_vector);
+    double IP = innerProduct(normalized, normalized);
 
     return (std::sqrt(IP / num_elements));
 
 }
 
-std::vector<std::vector<float>> normalizeData(std::vector<std::vector<float>> featuresMatrix) {
+std::vector<std::vector<double>> normalizeData(std::vector<std::vector<double>> featuresMatrix) {
     int num_cols = featuresMatrix[0].size();
     int num_rows = featuresMatrix.size();
-    std::vector<std::vector<float>> normalized_matrix(num_rows, std::vector<float>(num_cols, 0));
+    std::vector<std::vector<double>> normalized_matrix(num_rows, std::vector<double>(num_cols, 0));
 
     // for every column
     for (int j = 0; j < num_cols; j++) {
-        std::vector<float> col_to_normalize = getColumn(featuresMatrix, j);
-        std::vector<float> pre_normalized = subtractVectors(col_to_normalize, createVector(calculateMean(col_to_normalize), num_rows));
-        float col_STD = calculateSTD(col_to_normalize);
+        std::vector<double> col_to_normalize = getColumn(featuresMatrix, j);
+        std::vector<double> pre_normalized = subtractVectors(col_to_normalize, createVector(calculateMean(col_to_normalize), num_rows));
+        double col_STD = calculateSTD(col_to_normalize);
         if (col_STD == 0) {
             throw std::runtime_error("Standard deviation is zero, cannot normalize.");
         }
-        std::vector<float> normalized = divideVector(pre_normalized, col_STD);
+        std::vector<double> normalized = divideVector(pre_normalized, col_STD);
 
         updateColumn(normalized_matrix, normalized, j);
     }
